Move Gabby's pages to the view folder
I had mistakenly put them in components
