+++ frontend/src/router/index.js
@@ -1,6 +1,6 @@
 import {createRouter, createWebHistory} from 'vue-router'
-import OrgSearch from '../components/OrgSearch.vue'
-import ResultsPage from '../components/ResultsPage.vue'
+import OrgSearch from '../views/OrgSearch.vue'
+import ResultsPage from '../views/ResultsPage.vue'
 
 const routes = [
     {
